+++ pkarr/examples/resolve.rs
@@ -3,7 +3,6 @@
 //! run this example from the project root:
 //!     $ cargo run --example resolve <zbase32 encoded key>
 
-use tracing::Level;
 use tracing_subscriber;
 
 use std::{
@@ -27,8 +26,7 @@ struct Cli {
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
     tracing_subscriber::fmt()
-        .with_max_level(Level::DEBUG)
-        .with_env_filter("pkarr")
+        .with_env_filter("pkarr=info")
         .init();
 
     let cli = Cli::parse();